+++ src/extensions/options/optionsext.h
@@ -30,6 +30,7 @@
 #include "always.h"
 #include "extension.h"
 #include "options.h"
+#include "SDL3/SDL_surface.h"
 
 
 class CCINIClass;
@@ -81,4 +82,20 @@ class OptionsClassExtension final : public GlobalExtensionClass<OptionsClass>
         int KeyChatToAll1;
         int KeyChatToAll2;
         int KeyChatToAllies;
+
+        /**
+         *  Window size override.
+         */
+        int WindowWidth;
+        int WindowHeight;
+
+        /**
+         *  Scaling mode.
+         */
+        SDL_ScaleMode ScaleMode;
+
+        /**
+         *  Cursor scale factor.
+         */
+        int CursorScale;
 };